+++ lib/medic/test.ex
@@ -1,4 +1,6 @@
 defmodule Medic.Test do
+  # @related [test](/test/medic/test_test.exs)
+
   @moduledoc """
   Common list of tests.
 
@@ -9,8 +11,12 @@ defmodule Medic.Test do
       elixir -r .medic/require.exs -e "Medic.Test.run()" $*
 
   """
-  def run do
+  def run(opts \\ []) do
     Medic.Cmd.run!("Compiling", "mix", ["compile", "--force", "--warnings-as-errors"], env: [{"MIX_ENV", "test"}])
-    Medic.Cmd.run!("Running tests", "mix", ["test", "--color", "--warnings-as-errors"])
+    Medic.Cmd.run!("Running tests", "mix", ["test" | mix_test_options(opts)])
+  end
+
+  def mix_test_options(opts) do
+    ["--color", "--warnings-as-errors" | opts]
   end
 end